feat(preference): add preference module
